uint8 will handle cases where params are negative numbers

+++ definitions.go
@@ -13,12 +13,12 @@ const (
 var defaultConfig settings
 
 type setting struct {
-	MinLength         int `toml:"min_length" form:"min-length"`
-	SpecialCharacters int `toml:"special_characters" form:"special-chars"`
-	Numbers           int `toml:"numbers" form:"numbers"`
-	MinLowercase      int `toml:"min_lowercase" form:"min-lower"`
-	MinUppercase      int `toml:"min_uppercase" form:"min-upper"`
-	Results           int `toml:"results" form:"res"`
+	MinLength         uint8 `toml:"min_length" form:"min-length"`
+	SpecialCharacters uint8 `toml:"special_characters" form:"special-chars"`
+	Numbers           uint8 `toml:"numbers" form:"numbers"`
+	MinLowercase      uint8 `toml:"min_lowercase" form:"min-lower"`
+	MinUppercase      uint8 `toml:"min_uppercase" form:"min-upper"`
+	Results           uint8 `toml:"results" form:"res"`
 }
 
 type settings map[string]setting
